+++ R/import_readme.R
@@ -5,7 +5,7 @@
 
   # setup_docs() already created README.md if there is none, so if we don't find
   # any, it means the user has deleted it and we error
-  if (length(readme_files) == 0) {
+  if (!"README.md" %in% readme_files) {
     cli::cli_abort("README.md is mandatory.")
   }
 
@@ -44,30 +44,6 @@
     }
   }
 
-  # Render README if Rmd or qmd
-  # Do nothing if we only have a .md file
-  switch(
-    readme_type,
-
-    # rmd -> md
-    "rmd" = {
-        pre <- fs::path_join(c(src_dir, "altdoc/preamble_vignettes_rmd.yml"))
-        pre <- tryCatch(.readlines(pre), error = function(e) NULL)
-        .qmd2md(src_file, src_dir, preamble = pre)
-    },
-
-    # qmd -> md
-    "qmd" = {
-        pre <- fs::path_join(c(src_dir, "altdoc/preamble_vignettes_qmd.yml"))
-        pre <- tryCatch(.readlines(pre), warning = function(w) NULL, error = function(e) NULL)
-        # process in-place for use on Github
-        # TODO: preambles inserted in the README often break Quarto websites. It's
-        # not a big problem to omit the preamble, but it would be good to
-        # investigate this, because I am not sure what is going on -VAB
-        .qmd2md(src_file, src_dir)
-    }
-  )
-
   tar_file <- fs::path_join(c(tar_dir, "README.md"))
   src_file <- fs::path_join(c(src_dir, "README.md"))
   fs::file_copy(src_file, tar_file, overwrite = TRUE)
@@ -75,10 +51,15 @@
 
   # Add the index page which includes README.md
   if (tool == "quarto_website") {
-    writeLines(
-      enc2utf8("{{< include README.md >}}"),
-      fs::path_join(c(tar_dir, "index.md"))
-    )
+    if ("README.qmd" %in% readme_files) {
+      fs::file_copy(fs::path_join(c(src_dir, "README.qmd")), fs::path_join(c(tar_dir, "index.qmd")))
+    } else {
+      writeLines(
+        enc2utf8("{{< include README.md >}}"),
+        fs::path_join(c(tar_dir, "index.md"))
+      )
+      fs::file_copy(fs::path_join(c(src_dir, "README.md")), tar_dir)
+    }
   }
 
   tmp <- fs::path_join(c(src_dir, "README.markdown_strict_files"))
@@ -87,4 +68,7 @@
   }
   .update_freeze(src_dir, basename(src_file), successes = 1, fails = NULL, type = "README")
   cli::cli_alert_success("{.file README} imported.")
+  if ("README.qmd" %in% readme_files) {
+    cli::cli_alert("Altdoc does not render README.qmd automatically to markdown. Please ensure that your README.md file is in sync.")
+  }
 }